repository tests fixed
