--- src/test/kotlin/com/ticketera/ticketera/TicketeraApplicationTests.kt
@@ -1,17 +1,0 @@
-package com.ticketera.ticketera
-
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
-import org.springframework.boot.test.context.SpringBootTest
-import org.testcontainers.junit.jupiter.Testcontainers
-
-@Testcontainers
-@SpringBootTest
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-class TicketeraApplicationTests : ContainerTests() {
-
-    @Test
-    fun contextLoads() {
-    }
-}
