+++ src/test/kotlin/com/ticketera/ticketera/repositories/EventRepositoryTest.kt
@@ -1,6 +1,5 @@
 package com.ticketera.ticketera.repositories
 
-import com.ticketera.ticketera.ContainerTests
 import com.ticketera.ticketera.model.Event
 import com.ticketera.ticketera.model.Venue
 import org.assertj.core.api.Assertions.assertThat
@@ -10,17 +9,18 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager
+import org.springframework.test.context.DynamicPropertyRegistry
+import org.springframework.test.context.DynamicPropertySource
+import org.testcontainers.containers.PostgreSQLContainer
+import org.testcontainers.junit.jupiter.Container
 import org.testcontainers.junit.jupiter.Testcontainers
 import java.time.Instant
 import java.util.UUID
 
-@DataJpaTest
 @Testcontainers
-@AutoConfigureTestDatabase(
-    replace = AutoConfigureTestDatabase.Replace.NONE
-)
-class EventRepositoryTest : ContainerTests() {
-
+@DataJpaTest
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+final class EventRepositoryTest {
     @Autowired
     private lateinit var eventRepository: EventRepository
 
@@ -29,7 +29,6 @@ class EventRepositoryTest : ContainerTests() {
 
     private val venueId = UUID.randomUUID()
     private val eventId = UUID.randomUUID()
-
 
     @BeforeEach
     fun setup() {
@@ -75,6 +74,28 @@ class EventRepositoryTest : ContainerTests() {
         eventRepository.deleteById(eventId)
         assertThat(eventRepository.findById(venueId))
             .isEmpty
+    }
+
+    companion object {
+        @Container
+        val postgres = PostgreSQLContainer<Nothing>("postgres:16.4")
+            .apply {
+                withDatabaseName("ticketera")
+                withUsername("ticketera")
+                withPassword("dGVzdGRi")
+                withReuse(true)
+            }
+
+        @JvmStatic
+        @DynamicPropertySource
+        fun registerProperties(registry: DynamicPropertyRegistry) {
+            registry.add("spring.datasource.url", postgres::getJdbcUrl)
+            registry.add("spring.datasource.username", postgres::getUsername)
+            registry.add("spring.datasource.password", postgres::getPassword)
+            registry.add("spring.flyway.url", postgres::getJdbcUrl)
+            registry.add("spring.flyway.user", postgres::getUsername)
+            registry.add("spring.flyway.password", postgres::getPassword)
+        }
     }
 
 }
