+++ src/test/resources/application.yml
@@ -2,12 +2,17 @@ spring:
   application:
     name: ticketera
   datasource:
+    username:
+    password:
+    url:
     driverClassName: org.postgresql.Driver
     hikari:
       connection-timeout: 2000
       maximum-pool-size: 5
   flyway:
-    enabled: true
+    user:
+    password:
+    url:
 
   sql:
     init:
